requirements.txt
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
scikit-learn
python-dotenv
joblib